enhancement: renamed services to API
