+++ src/types/services/auth.ts
@@ -1,0 +1,3 @@
+export interface GetTokenResponse {
+  token: string;
+}
